Did assignment c3prj1_deck

+++ c3prj1_deck/deck.c
@@ -21,7 +21,7 @@ void print_hand(deck_t * hand) {
 
 int deck_contains(deck_t * d, card_t c) {
   card_t ** ptr = d->cards;
-  for(int i=0; i<(d->n_cards); i++) {
+  for (int i=0; i<(d->n_cards); i++) {
     if(suit_letter(**ptr) == suit_letter(c) && value_letter(**ptr) == value_letter(c)) {
       return 1;
     }
@@ -31,11 +31,11 @@ int deck_contains(deck_t * d, card_t c) {
 }
 
 void shuffle(deck_t * d) {
-  card_t ** card =d -> cards ;
+  card_t ** card = d->cards ;
   card_t * temp;
-  size_t n=d ->n_cards;
+  size_t n =d->n_cards;
   int rand;
-  for (size_t i=0 ;i< n/2 ;i++) {
+  for (int i=0; i<(n/2); i++) {
     rand = random()%n;
     temp = card[i];
     card[i] = card[rand];
@@ -44,6 +44,6 @@ void shuffle(deck_t * d) {
 }
 
 void assert_full_deck(deck_t * d) {
-  for (int i = 0; i < d->n_cards; i++)
-    assert(deck_contains(d, card_from_num(i)));  
+  for (int i=0; i<d->n_cards; i++)
+    assert(deck_contains(d, card_from_num(i)));
 }
